# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "websocket-service API 文档.",
        "title": "websocket-service RESTful API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/swagger/websocket-service",
    "paths": {
        "/sim-send/{pubsub}/{topic}": {
            "post": {
                "description": "模拟发送到前端消息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MessageSim"
                ],
                "summary": "模拟发送到Topic消息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "pubsub",
                        "name": "pubsub",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "web",
                        "name": "topic",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "模拟消息",
                        "name": "msg",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "20000": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "50000": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}